CocoaAction separation.
Since we want a "pure" repo for the Swift implementation and CocoaAction uses KVO, it should be separated.
